include/define.h: Pickup type defines added. include/player.h: Change mode (pickup mode) to 32 bits. server/c_object.c: command_pickup function modified to support all     the new pickup options. server/player.c: Fix for path_to_player to prevent crashes.  Add check     in move_player_attack just in case.  check_pick function modified to     support new pickup options.  do_some_living modified to eat flesh     items if player is starving and they don't have any food. server/spell_effect.c: cast_detection modified so that detect magic/curse     will not show objects beneath the floor.  Change so that detect_monster     and detect_evil spells copy the face of what they detected - gives better     appearance on client. socket/item.c: fix bug LookAt that ws checking blocked_los on wrong     offset. MSW 2001-08-28
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/crossfire@1264 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ include/player.h
@@ -78,7 +78,7 @@ typedef struct pl {
   unsigned char prev_fire_on;
   unsigned char prev_keycode; /* Previous command executed */
   unsigned char key_down;     /* Last move-key still held down */
-  uint16 mode;			/* Mode of player for pickup. */
+  uint32 mode;			/* Mode of player for pickup. */
   signed char digestion;	/* Any bonuses/penalties to digestion */
   signed char gen_hp;		/* Bonuses to regeneration speed of hp */
   signed char gen_sp;		/* Bonuses to regeneration speed of sp */
